Optimize loaded files count
