+++ www/js/terrain_generator/map.js
@@ -7,6 +7,19 @@ import {BIOMES} from "./biomes.js";
 export const SMOOTH_RAD         = 3;
 export const NO_SMOOTH_BIOMES   = [BIOMES.OCEAN.code, BIOMES.BEACH.code];
 
+export class MapCell {
+
+    constructor(value, humidity, equator, biome, block) {
+        this.value      = value;
+        this.value2     = value;
+        this.humidity   = Math.round(humidity * 100000) / 100000;
+        this.equator    = Math.round(equator * 100000) / 100000;
+        this.biome      = biome;
+        this.block      = block;
+    }
+
+}
+
 export class Map {
 
     static _cells;
