--- www/vendors/prismarine-physics/lib/math.js
@@ -1,5 +1,0 @@
-export class math {
-    static clamp(min, x, max) {
-        return Math.max(min, Math.min(x, max))
-    }
-}
